+++ providers/bing.js
@@ -5,6 +5,8 @@ module.exports.imagery = {
     'imagery',
     'world'
   ],
+  minZoom: 0,
+  maxZoom: 20,
   url: 'https://ecn.t{switch:0,1,2,3}.tiles.virtualearth.net/tiles/a{quadkey}.jpeg?g=5250',
   description: 'Tiles from Bing',
   attribution: 'Map data © Bing',