+++ rollup.config.js
@@ -1,3 +1,4 @@
+const path = require('path')
 const resolve = require('rollup-plugin-node-resolve')
 const commonjs = require('rollup-plugin-commonjs')
 const builtins = require('rollup-plugin-node-builtins')
@@ -5,8 +6,8 @@ const globals = require('rollup-plugin-node-globals')
 const json = require('rollup-plugin-json')
 
 module.exports = {
-  entry: 'index.js',
-  dest: 'docs/slippy-tile.min.js',
+  entry: path.join(__dirname, 'index.js'),
+  dest: path.join(__dirname, 'docs', 'slippy-tile.js'),
   format: 'umd',
   plugins: [
     json(),